# Reduce indexing and use switch statement in appendPrettyAny