--- a/pretty_test.go
+++ b/pretty_test.go
@@ -38,7 +38,7 @@ var example1 = []byte(`
 	],
 	"values2": {},
 	"values3": [],
-	"deep": {"deep":{"deep":[1,2,3,4,5]}}
+	"deep": {"deep":{"deep":[0,1,2,3,4,5,6,7,8,9]}}
 }
 `)