I noticed that some code includes string indexing for each conditional statements. I believe it's faster to reduce indexing to once in a loop and let the compiler optimize the switch statement (like [CL26770](https://go-review.googlesource.com/c/go/+/26770)).