@tidwall 2021-02-23
In this case using a switch could lead to slower runtime performance.
I specifically chose the ordering of the conditions in that loop.