@tidwall 2021-02-23
Before
```
BenchmarkPretty-16            	 1000000	      1033 ns/op	     720 B/op	       2 allocs/op
BenchmarkPrettySortKeys-16    	  594870	      2073 ns/op	    2848 B/op	      14 allocs/op
```

After
```
BenchmarkPretty-16            	 1000000	      1200 ns/op	     768 B/op	       2 allocs/op
BenchmarkPrettySortKeys-16    	  478676	      2129 ns/op	    2928 B/op	      14 allocs/op
```